feat : 위시리스트에 해당 상품 있을 때만 위시리스트에서 상품 삭제

+++ src/main/java/gift/service/OrderService.java
@@ -38,7 +38,9 @@ public class OrderService {
         Order order = new Order(null, option, quantity, LocalDateTime.now(),
             orderRequestDTO.message(), member);
         orderRepository.save(order);
-        wishlistService.removeWishlist(member.getEmail(), option.getProduct().getId());
+        if (wishlistService.isProductInWishlist(member.getEmail(), option.getProduct().getId())) {
+            wishlistService.removeWishlist(member.getEmail(), option.getProduct().getId());
+        }
         OrderResponseDTO orderResponseDTO = new OrderResponseDTO(order.getId(),
             order.getOption().getId(), order.getQuantity(), order.getOrderDateTime(),
             order.getMessage());
